+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Direct registration of Rates using  method: `RateLimiterRegistry#register(String, Rates)`.
 - Direct rate limiting using method: `RateLimiterRegistry#tryAcquire`.
 - Method `RateLimiterRegistry#isWithinLimit(K key)`.
+- Method `RateLimiterRegistry#deregister(String key)` and its variants for `Method` and `Class`.
 
 ### Changes